fix: harden reserve yield dashboard parsing
